samples: Fix run_all_samples to skip loader copy
If the loader is not present, don't try to copy it

+++ API-Samples/run_all_samples.ps1
@@ -23,7 +23,9 @@ function Exec
     }
 }
 
-Copy-Item -force ..\loader\$dPath\vulkan-1.dll $dPath
+if (Test-Path ..\loader\$dPath\vulkan-1.dll) {
+    Copy-Item -force ..\loader\$dPath\vulkan-1.dll $dPath
+}
 
 echo "Initialize Instance"
 Exec "01-init_instance"
